OncoPubMinerMonitor/requirements.txt
appdirs==1.4.4
APScheduler==3.9.1
attrs==21.2.0
bioc==1.3.7
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
ConfigArgParse==1.5.1
connection-pool==0.0.3
datrie==0.8.2
decorator==5.0.9
filelock==3.0.12
ftputil==5.0.1
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
importlib-resources==5.4.0
ipython-genutils==0.2.0
joblib==1.0.1
jsonlines==2.0.0
jsonschema==3.2.0
jupyter-core==4.7.1
lxml==4.6.3
markdown2==2.4.0
nbformat==5.1.3
nltk==3.6.2
plac==1.3.3
psutil==5.8.0
PuLP==2.4
pyaml==20.4.0
pyfiglet==0.8.post1
pymarc==4.1.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
ratelimiter==1.2.0.post0
regex==2021.8.28
requests==2.26.0
six==1.16.0
smart-open==5.1.0
smmap==4.0.0
snakemake==6.6.1
stopit==1.1.2
tabulate==0.8.9
toposort==1.6
tqdm==4.62.2
traitlets==4.3.3
typing-extensions==3.10.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
wget==3.2
wrapt==1.12.1
zipp==3.5.0
